Refactor imports and update exception handling
- Added ruff pre-commit hook for linting.
- Removed unused imports in various files to clean up the codebase.
- Updated exception handling to use the new exceptions module across multiple files.
- Simplified import statements for better readability and maintainability.
- Ensured consistent usage of TTException in place of util.TTException.
