+++ .pre-commit-config.yaml
@@ -10,6 +10,11 @@ repos:
       - id: black
         language_version: python3
         args: [--line-length=120]
+  - repo: https://github.com/astral-sh/ruff-pre-commit
+    rev: v0.8.4
+    hooks:
+      - id: ruff
+        args: [--select=F401, --fix]
   - repo: https://github.com/pre-commit/mirrors-clang-format
     rev: v18.1.7
     hooks:
