+++ docs/images/ttexalens_structure.py
@@ -9,7 +9,6 @@ pip install diagrams
 """
 
 from diagrams import Cluster, Diagram, Edge
-from diagrams.programming.flowchart import Action
 from diagrams.generic import Node
