+++ ttexalens/hardware/blackhole/baby_risc_debug.py
@@ -2,10 +2,9 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-from typing import Callable
 from ttexalens.hardware.baby_risc_debug import BabyRiscDebug
 from ttexalens.hardware.baby_risc_info import BabyRiscInfo
-from ttexalens.util import TTException
+from ttexalens.exceptions import TTException
 
 
 class BlackholeBabyRiscDebug(BabyRiscDebug):
