+++ ttexalens/tt_exalens_init.py
@@ -1,7 +1,6 @@
 # SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
 
 # SPDX-License-Identifier: Apache-2.0
-import os
 import atexit
 
 from ttexalens.umd_api import UmdApi, local_init
